dom/media/tests/mochitest/identity/test_setIdentityProviderWithErrors.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="../head.js"></script>
  <script type="application/javascript" src="../pc.js"></script>
  <script type="application/javascript" src="../templates.js"></script>
  <script type="application/javascript" src="identityevent.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
'use strict';
createHTML({
  title: 'Identity Provider returning errors is handled correctly'
});

var test;

runNetworkTest(function () {
  test = new PeerConnectionTest();
  test.setMediaConstraints([{audio: true}], [{audio: true}]);
  // first example generates a bad assertion that can be detected immediately
  test.setIdentityProvider(test.pcLocal, 'example.com', 'idp.js#bad-assert', 'nobody');
  // second generates a bad assertion, but that only fails to validate
  test.setIdentityProvider(test.pcRemote, 'example.com', 'idp.js#bad-validate', 'nobody');

  var localEvents = trapIdentityEvents(test.pcLocal._pc);
  var remoteEvents = trapIdentityEvents(test.pcRemote._pc);

  // if we get a peeridentity event, something is wrong
  function assertNoIdentity() {
    ok(!localEvents.peeridentity,
       'no peer identity event for local peer');
    ok(!remoteEvents.peeridentity,
       'no peer identity event for remote peer');
    throw new Error('got peeridentity, aborting test');
  }
  test.pcLocal._pc.onpeeridentity = assertNoIdentity;
  test.pcRemote._pc.onpeeridentity = assertNoIdentity;

  test.chain.append([
    [
      'CHECK_IDENTITY_EVENTS',
      function(t) {
        var checkEvents = () => {
          ok(localEvents.idpassertionerror,
             'local assertion generation should fail (idpassertionerror)');
          is(localEvents.idpassertionerror.idp,
             'example.com', 'event IdP is correct');
          is(localEvents.idpassertionerror.protocol,
             'idp.js#bad-assert', 'event IdP protocol is #bad-assert');
          ok(!remoteEvents.idpassertionerror,
             'remote assertion generation should succeed (idpassertionerror)');
          ok(!localEvents.identityresult,
             'local assertion generation should fail (identityresult)');
          ok(remoteEvents.identityresult,
             'remote assertion generation should succeed (identityresult)');

          ok(localEvents.idpvalidationerror, 'local fails to validate');
          is(localEvents.idpvalidationerror.idp, 'example.com',
             'event IdP is correct');
          is(localEvents.idpvalidationerror.protocol, 'idp.js#bad-validate',
             'event IdP protocol is #bad-validate');
          ok(!remoteEvents.idpvalidationerror,
             'remote doesn\'t even see an assertion');

          t.next();
        };

        // we actually have to wait on this because IdP validation is asynchronous
        if (localEvents.idpvalidationerror) {
          checkEvents();
        } else {
          // have to let the other event handlers have a chance to record success
          // before we run the checks that rely on that recording
          t.pcLocal._pc.onidpvalidationerror =
            e => window.setTimeout(checkEvents, 0);
        }
      }
    ],
    [
      'PEER_IDENTITY_IS_EMPTY',
      function(t) {
        ok(!t.pcLocal._pc.peerIdentity, 'local peerIdentity is not set');
        ok(!t.pcRemote._pc.peerIdentity, 'remote peerIdentity is not set');
        t.next();
      }
    ],
    [
      'ONLY_REMOTE_SDP_INCLUDES_IDENTITY_ASSERTION',
      function(t) {
        ok(!t.originalOffer.sdp.contains('a=identity'),
           'a=identity not contained in the offer SDP');
        ok(t.originalAnswer.sdp.contains('a=identity'),
           'a=identity is contained in the answer SDP');
        t.next();
      }
    ]
  ]);
  test.run();
});

</script>
</pre>
</body>
</html>
